+++ internal/middleware/storeid.go
@@ -1,4 +1,4 @@
-package storeid
+package middleware
 
 import (
 	"context"
@@ -7,8 +7,6 @@ import (
 	"go.opentelemetry.io/otel/trace"
 	"google.golang.org/grpc"
 )
-
-type ctxKey string
 
 const (
 	storeIDCtxKey ctxKey = "store-id-context-key"
@@ -42,43 +40,9 @@ func NewStoreIDInterceptor() grpc.UnaryServerInterceptor {
 	}
 }
 
-type wrappedServerStream struct {
-	grpc.ServerStream
-	wrappedContext context.Context
-}
-
-func newWrappedServerStream(stream grpc.ServerStream) *wrappedServerStream {
-	if existing, ok := stream.(*wrappedServerStream); ok {
-		return existing
-	}
-	return &wrappedServerStream{ServerStream: stream, wrappedContext: stream.Context()}
-}
-
-func (w *wrappedServerStream) Context() context.Context {
-	return w.wrappedContext
-}
-
-func (w *wrappedServerStream) RecvMsg(m interface{}) error {
-	if err := w.ServerStream.RecvMsg(m); err != nil {
-		return nil
-	}
-
-	if r, ok := m.(hasGetStoreID); ok {
-		storeID := r.GetStoreId()
-
-		// Add the storeID to the context
-		w.wrappedContext = context.WithValue(w.Context(), storeIDCtxKey, storeID)
-
-		// Add the storeID to the span
-		trace.SpanFromContext(w.Context()).SetAttributes(attribute.String(storeIDTraceKey, storeID))
-	}
-
-	return nil
-}
-
 // NewStreamingStoreIDInterceptor must come after the trace interceptor
 func NewStreamingStoreIDInterceptor() grpc.StreamServerInterceptor {
 	return func(srv interface{}, stream grpc.ServerStream, info *grpc.StreamServerInfo, handler grpc.StreamHandler) error {
-		return handler(srv, newWrappedServerStream(stream))
+		return handler(srv, newWrapperServerStream(stream))
 	}
 }
