+++ internal/middleware/storeid_test.go
@@ -1,4 +1,4 @@
-package storeid
+package middleware
 
 import (
 	"context"
